# docs/swagger.yml
openapi: 3.0.0
info:
  version: 0.0.0
  title: AlgoLearn
servers:
  - url: "http://localhost:8080"
paths:
...
  /solution:
    get:
      summary: "Get all solutions"
      operationId: listSolutions
      tags:
        - Solution
      responses:
        "200":
          description: "A list of all solutions"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Response_Solution"
    post:
      summary: "Create a new solution."
      operationId: createSolution
      tags:
        - Solution
      parameters:
        - $ref: "#/components/parameters/Auth_Token"
      requestBody:
        description: "Solution object to create"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Request_Solution"
      responses:
        "200":
          description: "Solution created"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response_Solution"
  /solution/{solutionId}:
    get:
      summary: "Get a solution by ID."
      operationId: getSolution
      tags:
        - Solution
      parameters:
        - name: solutionId
          in: path
          description: "ID of the solution"
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: "Details about a solution by ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response_Solution"
    patch:
      summary: "Modify an existing solution by ID."
      operationId: modifySolution
      tags:
        - Solution
      parameters:
        - name: solutionId
          in: path
          description: "ID of the solution"
          required: true
          schema:
            type: integer
        - $ref: "#/components/parameters/Auth_Token"
      requestBody:
        description: "Solution object to create"
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                solutionId:
                  type: integer
                solver:
                  type: integer
                problem:
                  type: integer
                code:
                  type: string
                passedTests:
                  type: integer
                compTime:
                  type: string
      responses:
        "200":
          description: "Solution modified"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response_Solution"
    put:
      summary: "Replace an existing solution by ID."
      operationId: replaceSolution
      tags:
        - Solution
      parameters:
        - name: solutionId
          in: path
          description: "ID of the solution"
          required: true
          schema:
            type: integer
        - $ref: "#/components/parameters/Auth_Token"
      requestBody:
        description: "Solution object to create"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Request_Solution"
      responses:
        "200":
          description: "Solution Replaced"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response_Solution"
    delete:
      summary: "Delete an existing solution"
      operationId: deleteSolution
      tags:
        - Solution
      parameters:
        - name: solutionId
          in: path
          description: "ID of solution"
          required: true
          schema:
            type: integer
        - $ref: "#/components/parameters/Auth_Token"
      responses:
        "200":
          description: "Solution Deleted"
  /solution/search:
    get:
      summary: "Search for a solution."
      operationId: searchSolution
      tags:
        - Solution
      parameters:
        - name: solutionId
          in: query
          description: "Part or all of the ID of the solution"
          required: false
          schema:
            type: integer
        - name: solver
          in: query
          description: "Part or all of the username of the user who submitted this solution"
          required: false
          schema:
            type: string
        - name: problem
          in: query
          description: "Part or all of the ID of the solution's problem"
          required: false
          schema:
            type: integer
        - name: page
          in: query
          description: "Page number of search results"
          required: false
          schema:
            type: integer
        - name: code
          in: query
          description: "Part or all of the source code in the solution"
          required: false
          schema:
            type: string
        - name: passedTests
          in: query
          description: "The number of test cases the solution passed"
          required: false
          schema:
            type: integer
        - name: compTime
          in: query
          description: "The amount of time required for the solution to run"
          required: false
          schema:
            type: string
      responses:
        "200":
          description: "List of solutions"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Response_Solution"
...
components:
  parameters:
    Auth_Token:
      name: Authorization
      in: header
      description: "Authorization token"
      requred: true
      schema:
        type: string
  schemas:
...
    Request_Solution:
      type: object
      required:
        - solver
        - problem
        - code
      properties:
        solutionId:
          type: integer
        solver:
          type: string
        problem:
          type: integer
        code:
          type: string
        tokens:
          type: string
    Response_Solution:
      type: object
      required:
        - solutionId
        - solver
        - problem
        - code
        - tokens
      properties:
        solutionId:
          type: integer
        solver:
          type: string
        problem:
          type: integer
        code:
          type: string
        passedTests:
          type: integer
        compTime:
          type: double
        tokens:
          type: string
...